+++ src/components/organization/projectOrganizations.tsx
@@ -8,13 +8,13 @@ import {
   ProjectOrganisationResponseDTO,
   OrganisationRoleType,
 } from "@team-golfslag/conflux-api-client/src/client.ts";
-import Organization from "@/components/organization.tsx";
+import Organization from "@/components/organization/organization";
 import { Label } from "@/components/ui/label.tsx";
-import EditOrganizationModal from "@/components/editOrganizationModal.tsx";
-import { Card, CardContent, CardHeader, CardTitle } from "./ui/card";
+import EditOrganizationModal from "@/components/contributor/editOrganizationModal";
+import { Card, CardContent, CardHeader, CardTitle } from "../ui/card";
 import { Edit, X, Archive, Building2 } from "lucide-react";
-import { Button } from "./ui/button";
-import { Badge } from "./ui/badge";
+import { Button } from "../ui/button";
+import { Badge } from "../ui/badge";
 
 export interface ProjectOrganizationsProps {
   isAdmin?: boolean;
