+++ src/components/overview/projectDescriptionSection.tsx
@@ -21,8 +21,8 @@ import {
   useEditableText,
   useTruncatableText,
 } from "@/hooks/useEditableContent";
-import { LanguageInput } from "@/components/languageInput";
-import { useLanguage } from "@/lib/LanguageContext";
+import { LanguageInput } from "@/components/ui/languageInput";
+import { useLanguage } from "@/lib/languageContext";
 import { Button } from "@/components/ui/button";
 import { Textarea } from "@/components/ui/textarea";
 import {
