+++ src/components/overview/projectTitleSection.tsx
@@ -8,8 +8,8 @@ import {
 } from "@team-golfslag/conflux-api-client/src/client";
 import { useApiMutation } from "@/hooks/useApiMutation";
 import { useEditableText } from "@/hooks/useEditableContent";
-import { LanguageInput } from "@/components/languageInput";
-import { useLanguage } from "@/lib/LanguageContext";
+import { LanguageInput } from "@/components/ui/languageInput";
+import { useLanguage } from "@/lib/languageContext";
 import { Button } from "@/components/ui/button";
 import { CardTitle } from "@/components/ui/card";
 import { Textarea } from "@/components/ui/textarea";
