+++ src/main.tsx
@@ -13,9 +13,9 @@ import { ApiClient } from "@team-golfslag/conflux-api-client/src/client";
 import config from "@/config.ts";
 import { ApiClientContext } from "@/lib/ApiClientContext.ts";
 import { SessionProvider } from "@/lib/SessionContext";
-import { LanguageProvider } from "@/lib/LanguageContext";
-import { BackendStatusProvider } from "@/lib/BackendStatusContext";
-import { BackendStatusWrapper } from "@/components/BackendStatusWrapper";
+import { LanguageProvider } from "@/lib/languageContext";
+import { BackendStatusProvider } from "@/lib/backendStatusContext";
+import { BackendStatusWrapper } from "@/components/backendStatusWrapper";
 
 /*
 This is the main entry point into the program
